+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Current nightly]
 
 ### Added
+- Added `-List` to `Get-PnPStoredCredential` to enumerate the names of the stored PnP credentials without requiring a credential name.
 - Added `Test-PnPSiteTemplate` to validate the schema, internal structure and packaged resources of a PnP site template without applying it. [#5444](https://github.com/pnp/powershell/pull/5444)
 - Added `-EnableSensitivityLabelForOneNote` and `-EnableSensitivityLabelForVideoFiles` parameters to `Set-PnPTenant`. [#5442](https://github.com/pnp/powershell/pull/5442)
 - Added `Test-PnPConnectionPermission` which compares the current connection's access token claims with the permission metadata of a PnP PowerShell cmdlet. [#5428](https://github.com/pnp/powershell/pull/5428)